requirements.txt
# Core frameworks
streamlit
langchain
langchain-openai

# Utilities
python-dotenv
requests
python-dateutil # For robust date parsing